Removed ID from payload view so it's not externally accessibled, moved payload structs to a new models package
