+++ glide.lock
@@ -1,5 +1,5 @@
 hash: 0cc9472c04bed080f1c8faebe4bb8337b8a7af04a2c6d71c43e03d4059e4b4f8
-updated: 2016-05-12T16:51:37.966701493+01:00
+updated: 2016-05-13T14:17:58.547720462+01:00
 imports:
 - name: github.com/boltdb/bolt
   version: c1c3bd7e847a231b2b1f9592fa86182a121ad734
