+++ import.go
@@ -13,6 +13,7 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"net/http"
 	"encoding/base64"
+	"github.com/SpectoLabs/hoverfly/models"
 )
 
 // Import is a function that based on input decides whether it is a local resource or whether
@@ -125,7 +126,7 @@ func isJSON(s string) bool {
 }
 
 // ImportPayloads - a function to save given payloads into the database.
-func (d *Hoverfly) ImportPayloads(payloads []PayloadView) error {
+func (d *Hoverfly) ImportPayloads(payloads []models.PayloadView) error {
 	if len(payloads) > 0 {
 		success := 0
 		failed := 0
@@ -159,13 +160,6 @@ func (d *Hoverfly) ImportPayloads(payloads []PayloadView) error {
 				}
 			}
 
-			// recalculating request hash and storing it in database
-			r := RequestContainer{Details: pl.Request, Minifier: d.MIN}
-			key := r.Hash()
-
-			// regenerating key
-			pl.ID = key
-
 			bts, err := pl.Encode()
 			if err != nil {
 				log.WithFields(log.Fields{
@@ -188,7 +182,7 @@ func (d *Hoverfly) ImportPayloads(payloads []PayloadView) error {
 					}).Error("failed to fire hook")
 				}
 
-				d.RequestCache.Set([]byte(key), bts)
+				d.RequestCache.Set([]byte(pl.Id()), bts)
 				if err == nil {
 					success++
 				} else {
